cofi-dashboard/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
mysql-connector-python==8.2.0
python-dotenv==1.0.0
sse-starlette==1.8.2
pydantic-settings==2.1.0
structlog==23.2.0
